project/backend/requirements.txt
a2a-sdk==0.3.7
ag-ui-protocol==0.1.9
apscheduler
audioop-lts==0.2.2
autopep8==2.3.2
azure-cognitiveservices-speech==1.46.0
azure-core==1.35.1
azure-cosmos==4.9.0
azure-identity==1.25.0
azure-keyvault-secrets==4.10.0
azure-search-documents==11.5.3
azure-storage-blob==12.26.0
bcrypt==5.0.0
browser-use==0.7.9
cloudpickle==3.1.1
contourpy==1.3.3
cryptography==46.0.1
Cython==3.1.4
debugpy==1.8.17
decorator==5.2.1
docker==7.1.0
fastapi==0.117.1
fastapi-azure-auth==5.2.0
google-genai==1.39.1
gremlinpython==3.7.4
itsdangerous==2.2.0
joblib==1.5.2
jmespath==1.0.1
jupyter
jwcrypto==1.5.6
kaleido==1.1.0
lxml==5.4.0
Markdown==3.9
markitdown[pdf,docx,pptx,xlsx,outlook]==0.1.3
matplotlib==3.10.6
mcp==1.15.0
mcp-use==1.3.10
msgraph-sdk==1.45.0
notion-client==2.5.0
numpy==2.3.3
oauthlib==3.3.1
openai==1.109.1
openai-agents==0.3.2
openpyxl==3.1.5
opentelemetry-sdk==1.37.0
opentelemetry-semantic-conventions==0.58b0
pandas==2.3.2
perplexityai==0.12.0
playwright==1.55.0
plotly==6.3.0
polars==1.33.1
psutil==7.1.0
pydub==0.25.1
pyodbc==5.2.0
pyperclip==1.11.0
pylint==3.3.8
PyMuPDF==1.26.4
pytest==8.4.2
pytest-playwright-asyncio==0.7.1
pytest-xdist==3.8.0
python-docx==1.2.0
python-dotenv==1.1.1
python-multipart==0.0.20
python-pptx==1.0.2
rapidfuzz==3.14.1
redis==6.4.0
requests==2.32.5
requests-oauthlib==2.0.0
rich==14.1.0
scikit-learn==1.7.2
scipy==1.16.2
seaborn==0.13.2
Sphinx==8.2.3
sphinx_rtd_theme==3.0.2
sphinx-toolbox==4.0.0
sphinxcontrib-napoleon==0.7
SQLAlchemy==2.0.43
streamlit==1.50.0
termcolor==3.1.0
tiktoken==0.11.0
trafilatura==2.0.0
uvicorn[standard]==0.37.0
websockets==15.0.1
